Refactor feeds bottom sheet expanded content
Using slots API pattern instead of providing multiple callbacks as params
